Implicitly resolve QData instances.
